minor: fix a sign-compare warning in memmatch()
This was reported by clang because sizeof is unsigned. Interestingly it
also increases compression speed by about 0.2%.

+++ src/slz.c
@@ -525,7 +525,7 @@ static inline long memmatch(const unsigned char *a, const unsigned char *b, long
 	unsigned long xor;
 
 	while (1) {
-		if (len + 2 * sizeof(long) > max) {
+		if ((long)(len + 2 * sizeof(long)) > max) {
 			while (len < max) {
 				if (a[len] != b[len])
 					break;
